Correct load errors
After a delete, the todo was not deleted from its category too. I corrected
that by adding a method to do that on the Todo class and call it when
the delete button is clicked.
Call the storage functions after creating a todo in a category, and after editing
or markComplete a todo tooo.
The category was not saved after a page reload so i correct that by defining if from the
raw data.